[Silver III] Title: 주유소, Time: 20 ms, Memory: 2800 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -4,9 +4,9 @@ using namespace std;
 
 int good[100010];
 int price[100010];
-int minprice;
-int savegood;
-int goodprice;
+long long minprice;
+long long savegood;
+long long goodprice;
 
 int main()
 {
